Addresses https://github.com/channelcat/sanic/issues/337. 

```
$ wrk -t 10 -c 100 -d 10 -s pipeline.lua http://localhost:8000
Running 10s test @ http://localhost:8000
  10 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    47.21ms   31.48ms 174.15ms   71.39%
    Req/Sec     1.27k   641.14     4.04k    67.24%
  126848 requests in 10.06s, 15.00MB read
Requests/sec:  12608.15
Transfer/sec:      1.49MB
```