@r0fls 2017-02-14
@kolanos apparently you added the `cleanup` method, which doesn't seem necessary -- was there any reason for this, aside from it _seeming_ like a good idea? Cleaning up the connection is preventing it from being reused, which is in turn violating HTTP 1.1