@zkanda 2017-02-14
Haha, just removing cleanup can indeed resolve the pipeline issue.
But it's 78% slower on my machine without it.

With cleanup:

```
➜  sanic git:(master) ✗ wrk -t 1 -c 100 -d 10 http://localhost:8000                        <<<
Running 10s test @ http://localhost:8000
  1 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.29ms  767.36us  12.59ms   61.52%
    Req/Sec    43.41k     2.80k   45.72k    89.00%
  432394 requests in 10.03s, 52.78MB read
Requests/sec:  43125.84
Transfer/sec:      5.26MB
```

Without cleanup:

```
➜  sanic git:(master) ✗ wrk -t 1 -c 100 -d 10 http://localhost:8000                        <<<
Running 10s test @ http://localhost:8000
  1 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    11.07ms    7.34ms  60.46ms   67.29%
    Req/Sec     9.43k     6.52k   43.01k    89.00%
  93852 requests in 10.03s, 11.46MB read
Requests/sec:   9359.23
Transfer/sec:      1.14MB
```