@kolanos 2017-02-14
@r0fls I can't remember the reasoning behind the cleanup method, so remove/alter as necessary.