@seemethere 2017-02-14
The cleanup method is so that keep-alive works as intended. After every request the variables that are set for the particular request should be cleaned up.

It's not a perfect implementation but it works for right now.